Remove coupling of sonar-plugin-api on Guava 10
in order to prepare upgrade to Guava 18 in core codebase
