+++ sonar-core/src/main/java/org/sonar/core/util/HttpsTrust.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.api.utils;
+package org.sonar.core.util;
 
 import javax.net.ssl.*;
 import java.net.HttpURLConnection;
