+++ sonar-core/src/test/java/org/sonar/core/util/HttpsTrustTest.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.api.utils;
+package org.sonar.core.util;
 
 import org.junit.Test;
