+++ sonar-plugin-api/pom.xml
@@ -182,12 +182,6 @@
       <artifactId>dbunit</artifactId>
       <scope>test</scope>
     </dependency>
-    <dependency>
-      <groupId>org.simpleframework</groupId>
-      <artifactId>simple</artifactId>
-      <version>4.1.21</version>
-      <scope>test</scope>
-    </dependency>
   </dependencies>
 
   <build>
