+++ sonar-plugin-api/src/main/java/org/sonar/api/resources/ResourceTypeTree.java
@@ -26,10 +26,10 @@ import com.google.common.collect.Collections2;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableListMultimap;
 import com.google.common.collect.ListMultimap;
-import com.google.common.collect.Lists;
 import org.sonar.api.ServerExtension;
 import org.sonar.api.task.TaskExtension;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
@@ -80,7 +80,7 @@ public class ResourceTypeTree implements TaskExtension, ServerExtension {
   }
 
   public static final class Builder {
-    private List<ResourceType> types = Lists.newArrayList();
+    private List<ResourceType> types = new ArrayList<>();
     private ListMultimap<String, String> relations = ArrayListMultimap.create();
     private ResourceType root;
